+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Rename again `CourseProductRelation` to `Offering`
 - Improve product display styling on course detail pages
 - Display discounted prices in catalog and course detail pages
+- Display certificate and normal discount prices for Syllabus Course Run
 
 ### Fixed